backend/pyproject.toml
[project]
name = "alpha-app"
version = "0.1.0"
description = "Alpha — the duck in the machine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "pendulum>=3.0.0",
    "aiohttp>=3.9.0",
    "httpx>=0.27.0",
    "claude-agent-sdk==0.1.50",
    "logfire[asyncpg]>=3.0.0",
    "redis>=5.0.0",
    "python-frontmatter>=1.1.0",
    "python-dotenv>=1.0.0",
    "Pillow>=11.0.0",
    "pillow-heif>=0.18.0",
    "APScheduler>=3.10.0,<4",
    "pyyaml>=6.0",
    "numpy>=2.4.3",
    "aiobotocore>=2.20.0",
    "tokenizers>=0.21.0",
    "html2text>=2024.2.26",
]

[project.scripts]
alpha = "alpha_app.main:run"
job = "alpha_app.jobs:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alpha_app"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that need live claude process or database",
]
filterwarnings = [
    "ignore::logfire._internal.config.LogfireNotConfiguredWarning",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
]
